// packages/web/src/shared/api/generated/model/analyticsCalendarResponseDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Finnn API
 * Backend API for Finnn finance tracking
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsCalendarDayDto } from "./analyticsCalendarDayDto";
import type { AnalyticsEffectiveRangeDto } from "./analyticsEffectiveRangeDto";

export interface AnalyticsCalendarResponseDto {
  baseCurrency: string;
  effectiveRange: AnalyticsEffectiveRangeDto;
  calendarDays: AnalyticsCalendarDayDto[];
}