// packages/web/src/shared/api/generated/model/analyticsLargestMovementDtoKind.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Finnn API
 * Backend API for Finnn finance tracking
 * OpenAPI spec version: 0.1.0
 */

export type AnalyticsLargestMovementDtoKind =
  (typeof AnalyticsLargestMovementDtoKind)[keyof typeof AnalyticsLargestMovementDtoKind];

export const AnalyticsLargestMovementDtoKind = {
  paymentTransaction: "paymentTransaction",
  transferTransaction: "transferTransaction",
  debtTransaction: "debtTransaction",
} as const;